% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_location}
\alias{extract_location}
\title{Extract locations from text (simple version)}
\usage{
extract_location(text)
}
\arguments{
\item{text}{}
}
\value{
list
}
\description{
Extract locations from text (simple version)
}
\examples{
extract_location(text)
}
