% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_air_temp.R
\name{get_air_temperature}
\alias{get_air_temperature}
\title{Get air temperature}
\usage{
get_air_temperature(text)
}
\arguments{
\item{text}{}
}
\value{
character
}
\description{
Get air temperature
}
\examples{
get_air_temperature(text)

}
