% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get lat-longs}
\usage{
get_coordinates(text)
}
\arguments{
\item{text}{}
}
\value{
matrix
}
\description{
Get lat-longs
}
\examples{
get_coordinates(text$text)

}
