% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Extract elevations from text}
\usage{
get_elevation(text)
}
\arguments{
\item{text}{}
}
\value{
list
}
\description{
Extract elevations from text
}
\examples{
get_elevation(text)
}
