% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wos_lite.R
\name{search_wos}
\alias{search_wos}
\title{Search Web of Science}
\usage{
search_wos(api_key = api_key, query, count = 100, first_record = 1)
}
\arguments{
\item{api_key}{(optional) the API key for accessing the Web of Science API}

\item{query}{the search query}

\item{count}{the number of results to return (default is 100)}

\item{first_record}{the index of the first result to return (default is 1)}
}
\value{
a list containing the search results and the total count of matched records
}
\description{
This function searches the Web of Science database using a query and returns the results.
}
\examples{
search_wos("my-api-key", "climate change")
}
