% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srUtils-package.R
\docType{package}
\name{srUtils-package}
\alias{srUtils}
\alias{srUtils-package}
\title{srUtils: Extract text for systematic maps}
\description{
Extract coordinates, elevations, locations, air temperature from pdfs or other texts
}
\author{
\strong{Maintainer}: Julian Flowers \email{jrf128@student.aru.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
