% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srutils_init.R, R/textract_init.R
\name{srutils_initialise}
\alias{srutils_initialise}
\title{Initialise srutils environment}
\usage{
srutils_initialise()

srutils_initialise()
}
\description{
This function creates a Python virtual environment named 'srutils' and installs the required Python packages using pip.
It also sets the RETICULATE_PYTHON environment variable to point to the Python binary within the virtual environment.
}
\examples{
textextract_initialize()
}
