% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_from_pdf.R
\name{tables_from_pdfs}
\alias{tables_from_pdfs}
\title{Tables from pdfs
Extracts tqbles from pdf files using the tabula-py Python package}
\usage{
tables_from_pdfs(pdf)
}
\arguments{
\item{pdf}{}
}
\value{
list
}
\description{
Tables from pdfs
Extracts tqbles from pdf files using the tabula-py Python package
}
\examples{
tables_from_pdfs
}
