% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extraction.R
\name{textextract}
\alias{textextract}
\title{Extract text
A wrapper for the python function \code{textract}}
\usage{
textextract(file)
}
\arguments{
\item{file}{}
}
\description{
Extract text
A wrapper for the python function \code{textract}
}
\examples{
textextract(file)
}
