% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_init.R
\name{textextract_initialize}
\alias{textextract_initialize}
\title{Initialize text extraction}
\usage{
textextract_initialize()
}
\description{
Initialize text extraction
}
\examples{
textextract_initialize()
}
