% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_toc.R
\name{unnest_toc}
\alias{unnest_toc}
\title{Unnest table of contents}
\usage{
unnest_toc(files)
}
\arguments{
\item{files}{}
}
\value{
data frame
}
\description{
Unnest table of contents
}
\examples{
unnest_toc(files)

}
